Remove Blockpay
It appears that Blockpay is Dead on Arrival. Ambitious project with near-zero development. Their website is barely functioning, their social media presence is non-existing, and for the last several months they've done nothing (considering their blog posts and roadmap). Even if a service providers wants to join in - they can't.

Gridcoin can be overwhelming as it is, I don't think we need to confuse new people more by referring to not-working/dead projects.