Added wallet version to homepage and blog
Added a new line in the downloads section to display the current wallet version 3.7.11.0.

Updated the blog to add current wallet version 3.7.11.0

Note: these will have to be updated manually for now, until we have something dynamic. Looks like we'll have to pull version number in dynamically from Git. @barton26 is looking into this with me.