Remove references to Whaleshares
This is following a 4-1 vote on the web-dev Slack channel.

We have not had a Gridcoin tagged post on Whaleshares in nearly a month and we are not seeing any real interaction with the userbase.

This change can be easily reverted if the platform sees more activity in the future.